+++ app/src/main/java/com/example/eatraw/ComparingPriceListActivity.kt
@@ -35,9 +35,9 @@ class ComparingPriceListActivity : AppCompatActivity() {
                 for (document in documents) {
                     val fishName = document.getString("f_name")
                     val count = (document["f_count"] as? Long)?.toInt() ?: 0
-                    val minCost = (document["f_min"] as? Double)?.toInt()
-                    val avgCost = (document["f_avg"] as? Double)?.toInt()
-                    val maxCost = (document["f_max"] as? Double)?.toInt()
+                    val minCost = (document["f_min"] as? Long)?.toInt()
+                    val avgCost = (document["f_avg"] as? Long)?.toInt()
+                    val maxCost = (document["f_max"] as? Long)?.toInt()
                     val fishImg = document.getString("f_img")
                     val season = document.getString("f_season")
 //                    val storageReference = FirebaseStorage.getInstance().reference
@@ -140,9 +140,9 @@ class ComparingPriceListActivity : AppCompatActivity() {
                 for (document in documents) {
                     val fishName = document.getString("f_name")
                     val count = document.getLong("f_count")?.toInt() ?: 0
-                    val minCost = (document["f_min"] as? Double)?.toInt()
-                    val avgCost = (document["f_avg"] as? Double)?.toInt()
-                    val maxCost = (document["f_max"] as? Double)?.toInt()
+                    val minCost = (document["f_min"] as? Long)?.toInt()
+                    val avgCost = (document["f_avg"] as? Long)?.toInt()
+                    val maxCost = (document["f_max"] as? Long)?.toInt()
                     val fishImg = document.getString("f_img")
                     val season = document.getString("f_season")
 //                    val storageReference = FirebaseStorage.getInstance().reference
